requirements.txt
gunicorn~=20.1.0
asgiref~=3.5.2
django-environ~=0.8.1
django==4
djangorestframework==3.13.1
drf-yasg==1.20.0
requests~=2.27.1
psycopg2~=2.9.3
channels~=3.0.4
websockets~=10.3
channels_redis~=3.4.0
redis~=4.3.4